chore: Update imports and add shallowRef to vite-config.ts
The code changes involve updating the imports in vite-config.ts to include the new shallowRef function from the vue package. This function is used in the vueConfig array to add the 'shallowRef' property. This change improves the reactivity of the code.

Note: The commit message follows the conventional format of "chore: <commit message>".