chore: Update npm dependencies and devDependencies
This commit updates the npm dependencies and devDependencies in the package.json file. It includes updates to the following packages:
- vue-toastification: updated to version 2.0.0-rc.5
- sass: updated to version 1.77.8
- typescript: updated to version 5.5.3
- unplugin-auto-import: updated to version 0.17.8

These updates ensure that the project is using the latest stable versions of these packages, which may include bug fixes, performance improvements, and new features.

Note: The commit message follows the conventional format of "chore: <commit message>".